+++ binless/src/gfl_graph_fl.c
@@ -2,7 +2,7 @@
 #include "gfl_graph_fl.h"
 
 int graph_fused_lasso (int n, double *y,
-                        int ntrails, int *trails, int *breakpoints,
+                        int ntrails, unsigned *trails, unsigned *breakpoints,
                         double lam, double *alpha, double inflate,
                         int maxsteps, double converge,
                         double *beta)
@@ -12,7 +12,7 @@ int graph_fused_lasso (int n, double *y,
 
 
 int graph_fused_lasso_weight (int n, double *y, double *w,
-                        int ntrails, int *trails, int *breakpoints,
+                        int ntrails, unsigned *trails, unsigned *breakpoints,
                         double lam, double *alpha, double inflate,
                         int maxsteps, double converge,
                         double *beta)
@@ -42,7 +42,7 @@ int graph_fused_lasso_weight (int n, double *y, double *w,
 }
 
 int graph_fused_lasso_warm (int n, double *y,
-                        int ntrails, int *trails, int *breakpoints,
+                        int ntrails, unsigned *trails, unsigned *breakpoints,
                         double lam, double *alpha, double inflate,
                         int maxsteps, double converge,
                         double *beta, double *z, double *u)